Incredibly pedantic change proposal: Use 'npm' lower-case, as that is what npm wants
I'm not 100% sure why I'm making the pull request, but npm likes to be lower-case. I know I could have spent this time learning about something new (like datalog) or perhaps sleeping or writing an epic poem, or perhaps even getting a job. But instead, I changed the word NPM to npm four times, and then wrote this bit of gibberish to tell you about it.

According the the faq, npm should always be lower-case, except in man pages. This readme isn't really a man page. https://docs.npmjs.com/misc/faq#is-it-npm-or-npm-or-npm